OSIS-607 Write unittest for json generation
